+++ graylog2-web-interface/src/components/streams/StreamDetails/output-filter/FilterDeleteButton.tsx
@@ -20,10 +20,8 @@ import { useQueryClient } from '@tanstack/react-query';
 
 import { Button } from 'components/bootstrap';
 import { ConfirmDialog, Icon } from 'components/common';
-
-import type { StreamOutputFilterRule } from './Types';
-
-import useStreamOutputRuleMutation from '../../hooks/useStreamOutputRuleMutation';
+import type { StreamOutputFilterRule } from 'components/streams/StreamDetails/output-filter/Types';
+import useStreamOutputRuleMutation from 'components/streams/hooks/useStreamOutputRuleMutation';
 
 type Props ={
   streamId: string,
