+++ graylog2-web-interface/src/components/streams/StreamDetails/output-filter/Types.ts
