+++ graylog2-web-interface/src/components/streams/StreamDetails/routing-destination/OutputItem.tsx
@@ -19,11 +19,9 @@ import styled, { css } from 'styled-components';
 
 import type { Output } from 'stores/outputs/OutputsStore';
 import type { ConfigurationFormData } from 'components/configurationforms';
-
-import EditOutputButton from './EditOutputButton';
-import RemoveOutputButton from './RemoveOutputButton';
-
-import type { AvailableOutputRequestedConfiguration } from '../../useAvailableOutputTypes';
+import type { AvailableOutputRequestedConfiguration } from 'components/streams/useAvailableOutputTypes';
+import EditOutputButton from 'components/streams/StreamDetails/routing-destination/EditOutputButton';
+import RemoveOutputButton from 'components/streams/StreamDetails/routing-destination/RemoveOutputButton';
 
 type Props = {
   output: Output,
